Code refactoring for client package
Code refactoring for client package

Code refactoring for client package
