+++ packages/client/src/property/hideProperty.ts
@@ -1,0 +1,16 @@
+import { hashProperty, Property, PropertyType } from '@0auth/message';
+
+export function hideProperty(
+  properties: Property[],
+  hideNames: string[],
+): Property[] {
+  return properties.map((property) =>
+    hideNames.includes(property.key)
+      ? {
+          type: PropertyType.Hash,
+          key: property.key,
+          value: hashProperty(property),
+        }
+      : property,
+  );
+}
