+++ packages/client/src/property/index.ts
@@ -1,0 +1,1 @@
+export * from './hideProperty';
