+++ packages/client/src/utils/decryptMessage.ts
@@ -1,0 +1,5 @@
+import { AES, enc } from 'crypto-js';
+
+export function decryptMessage(encryptedMessage: string, key: string): string {
+  return AES.decrypt(encryptedMessage, key).toString(enc.Utf8);
+}
