+++ packages/client/src/utils/encryptMessage.ts
@@ -1,0 +1,5 @@
+import { AES } from 'crypto-js';
+
+export function encryptMessage(message: string, key: string): string {
+  return AES.encrypt(message, key).toString();
+}
